+++ web/ui/module/codemirror-promql/src/grammar/test/test-promql.test.ts
@@ -1,5 +1,5 @@
 import { parser } from '../parser';
-import { fileTests } from 'lezer-generator/dist/test';
+import { fileTests } from '@lezer/generator/dist/test';
 
 import * as fs from 'fs';
 import * as path from 'path';